requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
attrs==24.2.0
av==13.1.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==4.5.0
decorator==5.1.1
distro==1.9.0
docopt==0.6.2
executing==2.1.0
faiss-cpu==1.8.0.post1
faster-whisper==1.1.0
filelock==3.16.1
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.10.0
groq==0.12.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
ipython==8.29.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.7.0
joblib==1.4.2
Js2Py==0.74
jsonpatch==1.33
jsonpickle==4.0.0
jsonpointer==3.0.0
langchain==0.3.7
langchain-core==0.3.15
langchain-groq==0.2.1
langchain-ollama==0.2.0
langchain-openai==0.2.5
langchain-text-splitters==0.3.2
langsmith==0.1.139
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
moviepy==2.0.0
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==1.26.4
ollama==0.3.3
onnxruntime==1.20.1
openai==1.53.0
opencv-python==4.10.0.84
orjson==3.10.11
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
pip==24.3.1
proglog==0.1.10
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.3
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyjsparser==2.7.1
PyPrind==2.11.3
pySmartDL==1.3.4
python-dotenv==1.0.1
pyvis==0.3.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.3.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
stack-data==0.6.3
sympy==1.13.3
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.3
tqdm==4.66.6
traitlets==5.14.3
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
wcwidth==0.2.13
yarl==1.17.1
